ENH: FCTMF - No longer need to "define" a suite.
I realized that we could scope out a new block and define the function as a
variable and call that function all in one macro rather than breaking it into
two macros.

This reduces,

    FCTMF_SUITE_DEF(x);
    FCT_BGN() {
       FCTMF_SUITE_CALL(x);
    } FCT_END();

to the following,

    FCT_BGN() {
       FCTMF_SUITE_CALL(x);
    } FCT_END();

assuming that a FCTMF_SUITE is defined in another object file within the same
executable.
----------------------------------------------------------
e --include-merges or -n0 to see merged revisions.