FIX: fct - const char warnings with g++.
Fixed most of the bad "const char*" usage. However, there is one that I am
leaving for now as a "bad thing". The casting to char* in the initialization
structure.

I would like to grow a specialized initialization structure that doesn't
suffer from this problem, and provides the "right solution" in terms of
managing data responsibilities.