Start at the replace %envvar% code (fctstr_envrepl)
Probably has been written for the million'th time in the history of c
programming.

This code will serve as a backbone to the entire test data (iss2)
system.

By using %vars%, we can introduce lots of internal ones like
"%file_dir%", which allows for relative searches to the test source
file.