Fix fct_clp__optval2, now returns default string.
The default string was not being returned if the command line option was
not present when the program was being executed.

So for instance if you had a --logger option and you always wanted it to
default to "default" value, it would return NULL if the user didn't
specify --logger=<any-value> at the command prompt. This wasn't the
intended effect.

Unit tests where updated to catch this problem. Original unit tests
where using NULL as a default value, now additional tests use a string.