Fix warnings/errors from gcc compiler.
GCC compiler detects that you are passing NULL to a strlen() function
and rightly warns you about it. Thanks GCC! Problem is that my macros
for checking if a string is empty will inadvertently expand to NULL
within a strlen(), even though I took care to escape calling strlen when
the supplied string is NULL.

The fix involved adding two little functions rather than relying on a
macro expansion.

It is only an error because the build is configured to treat warnings as
errors.