Adds duration to test objects
Rather than having different logger implementations being responsible
for timing tests, this relatively common operation has been added to the
test object itself (see fct_test__duration()). Obviously this call is
only feasible when the test is all done.

There is also a fct_ts__duration that at the moment is only a summation
of all the sub-tests.

This changed allowed for some nice simplifications in the junit logger.