Fix assert(expr && "str")/assert("str") warnings on Intel 11.1.
Intel spits out warning #279: controlling expression is constant when
fct.h is included.  Further, because of FCT_END(), user source files emit
the same warning.  Fix is to replace assert -> FCT_ASSERT within fct.h
and then using _Pragma to disable to offending warning around each call
to assert.

Changes pass on GCC 4.2.4, GCC 4.3.4.
Warnings suppressed on Intel 11.1 20091130.

Note that Intel 11.1 seems to have trouble with test_big.c but the
changes in this patch do not seem to be the culprit.  Relevant message
emitted is:
    "An internal threshold was exceeded for routine main and
    optimization level may be reduced. See
    http://software.intel.com/en-us/articles/internal-threshold-was-exceeded
    for more information and advice."